dispatch action only when Create clicked
