+++ package.json
@@ -25,7 +25,8 @@
     "react-scripts": "5.0.0",
     "sass": "^1.49.9",
     "styled-components": "^5.3.3",
-    "typescript": "~4.1.5"
+    "typescript": "~4.1.5",
+    "uuid": "^8.3.2"
   },
   "scripts": {
     "start": "react-scripts start",
